(#7413)  Eliminate a pair of warnings.

+++ src/condor_utils/submit_utils.cpp
@@ -6370,14 +6370,14 @@ int SubmitHash::SetRequirements()
 		int convertedLength = 0;
 		bool setCUDAVersion = false;
 		if( sscanf( requiredCudaVersion.c_str(), "%u.%u%n", & major, & minor, & convertedLength ) == 2 ) {
-			if( convertedLength == requiredCudaVersion.length() ) {
+			if( (unsigned)convertedLength == requiredCudaVersion.length() ) {
 				long long int rcv = (major * 1000) + (minor % 100);
 				AssignJobVal(ATTR_CUDA_VERSION, rcv);
 				answer += "&& " ATTR_CUDA_VERSION " <= TARGET.MaxSupportedCUDAVersion";
 				setCUDAVersion = true;
 			}
 		} else if( sscanf( requiredCudaVersion.c_str(), "%u%n", & major, & convertedLength ) == 1 ) {
-			if( convertedLength == requiredCudaVersion.length() ) {
+			if( (unsigned)convertedLength == requiredCudaVersion.length() ) {
 				long long int rcv = major;
 				if( major < 1000 ) { rcv = major * 1000; }
 				AssignJobVal(ATTR_CUDA_VERSION, rcv);
